Add AVR_keywords library.properties
Now that the folder structure of the library is in 1.5 format Arduino
IDE 1.6.6 and 1.6.7 gives an Invalid Library warning for the
AVR_keywords library and the keyword highlighting doesn't work with
those IDE versions. The library.properties file is the same as the one
added in
https://github.com/MCUdude/MightyCore/commit/bb7c0cffd68b49b4f072d4fe81761cad84501d13
except I have changed the category to a valid value to fix the warning
and added the url field, without which the library is considered
invalid.