Add preprocessor recipes
The default arduino-builder recipe.preproc.includes and
recipe.preproc.macros don't have the -mmcu option which causes the MCU
specific macro to not be defined. For more information see
https://github.com/arduino/arduino-builder/issues/106. These are the
recipes from the Arduino AVR Boards 1.6.9 version included with Arduino
IDE 1.6.7 with a default value of null set for preprocessed_file_path to
allow for compatibility with Arduino IDE 1.6.6.