Add copies of all Arduino AVR Boards programmers
The platform.txt of the selected programmer is used for Sketch > Upload
Using Programmer. This means that the MightyCore avrdude.conf is not
used for Upload Using Programmer when any programmer not in the
MightyCore programmers.txt is selected which causes failure(eg.,
`avrdude: AVR Part "atmega324a" not found.`) for any of the boards that
require the added part definitions. This issue is fixed by selecting the
MightyCore copy added in this commit from Tools > Programmer.

These programmers are copied from Arduino AVR Boards 1.6.9 with only the
name values modified to identify them as the MightyCore copies. I left
off the Atmel STK500 development board programmer because it is
identical to the STK500 as ISP programmer already included with
MightyCore.