Improve recipe.preproc.macros IDE 1.6.6 compatibility
With Arduino IDE 1.6.6(only) the previous default preprocessed_file_path value was causing a file
named null to be created in the root of the Arduino IDE installation
folder which can be avoided on Windows by using the filename nul. A file
named nul will still be created on other OSs but now will be placed in
the temporary build folder which will avoid possible problems if the
user doesn't have write permission to the IDE installation folder.