Added new clock speed settings for the 1284, 644, and 164
Most AVR chips provide the ability to emit the processor's clock cycle
on a specific GPIO (the name provided is CLKO). This is done by changing
some of the fuses at bootloader programming time. The upside is that the
same clock that powers the microcontroller can be provided to other
devices in your project.

The downside is that you lose the use of the CLKO pin as a GPIO.

I have exposed this option as a different speed grade for 16 and 20 MHz
for the 1284, 644, and 164 Atmega chips. This ability isn't bound to
just 16 and 20MHz speeds but I have no need for the other speeds in my own projects.

I also only exposed this for the 1284, 644, and 164 because those are
the chips actually own and can easily test. I did not want to expose the
option on chips that I myself do not own and cause a problem for someone
else.